requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
astroid==2.12.5
atomicwrites==1.4.1
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
colorama==0.4.5
coverage==6.4.2
crashtest==0.3.1
dateutils==0.6.12
dill==0.3.5.1
distlib==0.3.6
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
dulwich==0.20.46
factory-boy==3.2.1
Faker==15.1.1
filelock==3.8.0
Flask==2.1.3
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
marshmallow==3.14.1
marshmallow-dataclass==8.5.8
mccabe==0.7.0
more-itertools==8.14.0
msgpack==1.0.4
mypy==0.910
mypy-extensions==0.4.3
npm==0.1.1
optional-django==0.1.0
packaging==21.3
pexpect==4.8.0
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.1.2
psycopg2-binary~=2.9.5
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pylev==1.4.0
pylint==2.15.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-env==0.6.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing-inspect==0.8.0
typing_extensions==4.3.0
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.16.5
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
